Client/requirements.txt
Appium-Python-Client==0.31
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.6
cffi==1.12.2
colorama==0.4.1
cryptography==2.6.1
html-testRunner==1.1.2
idna==2.8
Jinja2==2.9.5
MarkupSafe==1.1.1
more-itertools==4.3.0
paramiko==2.4.2
pathlib2==2.3.3
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
